+++ infrastructure/modules/monitoring/metrics-server/outputs.tf
@@ -1,0 +1,3 @@
+output "metrics_server" {
+  value = {}
+}