+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Added
 * Add HPA on ArmoniK control plane based on CPU and RAM utilization metrics
 * Add node selector on all deployments (ArmoniK and Helm charts for infrastructure)
 * Add logs of metrics exporter, MongoDB, NGINX, and Keda in CloudWatch via Fluent-bit
+* Add metrics server on EKS to use the HPA for ArmoniK control plane
 
 Changed
 -